motorolamodem: Fix configure warnings and header includes
We have a mismatch between MOTMODEM and MOTOROLAMODEM and won't be
able to properly configure motorolamodem. Let's standardize on
using MOTORLAMODEM for configure as that's the modem name as
suggested by Merlijn Wajer <merlijn@wizzup.org>.

We need to include headers from the users to avoid including them
multiple times, let's sort that out in preparation for adding sms
support.