gatsyntax: Add handling for Motorola custom line header
On motorola modems we have U1234 packet ID in the beginning of each
response line where the number is the packet ID. We can just ignore it
for now and handle it with a wrapper for g_at_syntax_new_gsm_permissive().

Note that we also add handling for firmware that terminates lines with
'\n' instead of '\r'. This could possibly be added to the handling for
GSM_PERMISSIVE_STATE_RESPONSE in gsm_permissive_feed(), but it's unclear
if that might break other modems.