motorolamodem: Add network strength notifications using qmimodem
On motorolamodem, the notifications are hardcoded to always go to the
/dev/gsmtty* TS 27.010 serial devices. Let's just add network strength
parsing from the serial, and update ofono status on behalf of qmimodem.

We already get a percentage in the RSSI status from /dev/gsmtty1. The
network strength parsing is based on an earlier patch from Pavel Machek
<pavel@ucw.cz>.

Note that the notifications can be disabled when the LCD is blanked and
there are no calls happening. This can be done  with:

$ printf "U1234AT+SCRN=0\r" > /dev/gsmtty1

And the notifications can be enabled again with:

$ printf "U1234AT+SCRN=1\r" > /dev/gsmtty1