qmimodem: Update WDA data format for network interfaces != qmi_wwan
The network interface for the qmimodem may be provided by another
driver, e.g. bam-dmux for the integrated modem in Qualcomm SoCs or IPA.
Those do not support changing the expected data format, they just support
using the modem in Raw-IP mode. However, it may be necessary to change
the WDA data format of the modem to make that work.

To implement that:
  1. Assume the network interface expects Raw-IP if it is not exposed
     via USB (and therefore managed by qmi_wwan).
  2. Fall back to changing the modem data format with "WDA Set Data Format"
     if changing the expected data format fails.