rmnet: Default created interfaces to MTU of 1400
By default the network interface is being created with 16380 byte MTU,
which is excessive for cellular links.  Typically, cellular links use a
maximum MTU of around 1430.  Set the MTU to something more reasonable by
default.

Unfortunately, the kernel does not support setting the MTU value in the
original NEWLINK message.  It must be set separately in a different
transaction.  Add support for this and make sure that cancellation logic
still works.

The logic in update_new_link_ifindex() is updated to make it more
defensive.  Since two rtnl messages are sent (RTM_NEWLINK to create the
device followed by RTM_SETLINK to set the MTU), there are now two
RTM_NEWLINK events that are triggered for each item in the rmnet
request.  Since 'current' is incremented after RTM_NEWLINK request
succeeds, there is a possibility of buffer overflow.