qmi: netreg: Fix RAT detection
It is not clear why extract_ss_info is looping over all ss->radio_if_count
interfaces, given that the result of that loops is that it always gets the
last RAT. However, if there is more than one active interface, like:

qmicli  -d  /dev/cdc-wdm1 --nas-get-serving-system
[/dev/cdc-wdm1] Successfully got serving system:
	Registration state: 'registered'
	CS: 'attached'
	PS: 'detached'
	Selected network: '3gpp'
	Radio interfaces: '2'
		[0]: 'umts'
		[1]: 'cdma-1xevdo'
...

the current logic will set tech to invalid (-1).

Fix that by looping backwards until a valid tech is set. That way the logic
is kept intact, besides that invalid RATs are filtered